requirements.txt
httpx
ipython
langchain
langchain-openai
langgraph
pandas
python-dotenv